+++ app/Http/Controllers/CommentaireController.php
@@ -3,9 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Commentaire;
+use App\Models\Publication;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StoreCommentaireRequest;
 use App\Http\Requests\UpdateCommentaireRequest;
-use App\Models\Publication;
 
 class CommentaireController extends Controller
 {
@@ -14,6 +15,12 @@ class CommentaireController extends Controller
      */
     public function index()
     {
+        $user = Auth::user();
+        if(!$user->can('list_commentaire'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $commentaires = Commentaire::with('publication')->get();
         return view('admin.commentaires.index', compact('commentaires'));
     }
@@ -23,6 +30,12 @@ class CommentaireController extends Controller
      */
     public function create()
     {
+        $user = Auth::user();
+        if(!$user->can('create_commentaire'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         return view('admin.commentaires.create', [
             'commentaire' => new Commentaire(),
             'publications' => Publication::pluck('type', 'id'),
@@ -34,12 +47,24 @@ class CommentaireController extends Controller
      */
     public function store(StoreCommentaireRequest $request)
     {
+        $user = Auth::user();
+        if(!$user->can('create_commentaire'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $commentaire = Commentaire::create($request->validated());
         return to_route('commentaire.index')->with('status', 'Ajout du commentaire effectué avec succès');
     }
 
     public function show(Commentaire $commentaire)
     {
+        $user = Auth::user();
+        if(!$user->can('show_commentaire'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         return view('admin.commentaires.show', compact('commentaire')); 
     }
 
@@ -48,6 +73,12 @@ class CommentaireController extends Controller
      */
     public function edit(Commentaire $commentaire)
     {
+        $user = Auth::user();
+        if(!$user->can('edit_commentaire'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         return view('admin.commentaires.edit', [
             'commentaire' => $commentaire,
             'publications' => Publication::pluck('type', 'id'),
@@ -59,6 +90,12 @@ class CommentaireController extends Controller
      */
     public function update(UpdateCommentaireRequest $request, Commentaire $commentaire)
     {
+        $user = Auth::user();
+        if(!$user->can('edit_commentaire'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $commentaire->update($request->validated());
         return to_route('commentaire.index')->with('status', 'Modification du commentaire effectuée  avec succès');
     }
@@ -68,6 +105,12 @@ class CommentaireController extends Controller
      */
     public function destroy(Commentaire $commentaire)
     {
+        $user = Auth::user();
+        if(!$user->can('delete_commentaire'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $commentaire->delete();
         return to_route('commentaire.index')->with('status', 'suppression du commentaire effectuée avec succès');
     }
