+++ app/Http/Controllers/HarcelementController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Models\Harcelement;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StoreHarcelementRequest;
 use App\Http\Requests\UpdateHarcelementRequest;
 
@@ -14,6 +16,13 @@ class HarcelementController extends Controller
      */
     public function index()
     {
+        $user = Auth::user();
+        if(!$user->can('list_harcelement'))
+        {
+            // dd($user);
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+        
         $harcelements = Harcelement::all();
         return view('admin.harcelements.index', compact('harcelements'));
     }
@@ -23,6 +32,12 @@ class HarcelementController extends Controller
      */
     public function create()
     {
+        $user = Auth::user();
+        if(!$user->can('create_harcelement'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         return view('admin.harcelements.create', [
             'harcelement' => new Harcelement(),
         ]);
@@ -33,6 +48,12 @@ class HarcelementController extends Controller
      */
     public function store(StoreHarcelementRequest $request)
     {
+        $user = Auth::user();
+        if(!$user->can('create_harcelement'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $harcelement = Harcelement::create($request->validated());
         return to_route('harcelement.index')->with('status', 'Ajout effectué avec succès');
     }
@@ -48,6 +69,12 @@ class HarcelementController extends Controller
      */
     public function edit(Harcelement $harcelement)
     {
+        $user = Auth::user();
+        if(!$user->can('edit_harcelement'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         return view('admin.harcelements.edit', [
             'harcelement' => $harcelement,
         ]);
@@ -58,6 +85,12 @@ class HarcelementController extends Controller
      */
     public function update(UpdateHarcelementRequest $request, Harcelement $harcelement)
     {
+        $user = Auth::user();
+        if(!$user->can('edit_harcelement'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $harcelement->update($request->validated());
         return to_route('harcelement.index')->with('status', 'Modification effectuée avec succès');
     }
@@ -67,6 +100,12 @@ class HarcelementController extends Controller
      */
     public function destroy(Harcelement $harcelement)
     {
+        $user = Auth::user();
+        if(!$user->can('delete_harcelement'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $harcelement->delete();
         return to_route('harcelement.index')->with('status', 'Suppression effectuée avec succès');
     }
