+++ app/Http/Controllers/TutorielController.php
@@ -3,9 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Tutoriel;
+use App\Models\Harcelement;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\StoreTutorielRequest;
 use App\Http\Requests\UpdateTutorielRequest;
-use App\Models\Harcelement;
 
 class TutorielController extends Controller
 {
@@ -14,6 +15,13 @@ class TutorielController extends Controller
      */
     public function index()
     {
+        $user = Auth::user();
+        if(!$user->can('list_tutoriel')) 
+        {
+            dd($user);
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $tutoriels = Tutoriel::with('harcelement')->get();
         return view('admin.tutoriels.index', compact('tutoriels'));
     }
@@ -23,6 +31,12 @@ class TutorielController extends Controller
      */
     public function create()
     {
+        $user = Auth::user();
+        if(!$user->can('create_tutoriel'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         return view('admin.tutoriels.create', [
             'tutoriel' => new Tutoriel(),
             'harcelements' => Harcelement::pluck('type', 'id'),
@@ -34,12 +48,24 @@ class TutorielController extends Controller
      */
     public function store(StoreTutorielRequest $request)
     {
+        $user = Auth::user();
+        if(!$user->can('create_tutoriel'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $tutoriel = Tutoriel::create($request->validated());
         return to_route('tutoriel.index')->with('status', 'Ajout effectué avec succès');
     }
 
     public function show(Tutoriel $tutoriel) 
     {
+        $user = Auth::user();
+        if(!$user->can('show_tutoriel'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         return view('admin.tutoriels.show', compact('tutoriel'));  
     }
 
@@ -48,6 +74,12 @@ class TutorielController extends Controller
      */
     public function edit(Tutoriel $tutoriel)
     {
+        $user = Auth::user();
+        if(!$user->can('edit_tutoriel'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         return view('admin.tutoriels.edit', [
             'tutoriel' => $tutoriel,
             'harcelements' => Harcelement::pluck('type', 'id'),
@@ -59,6 +91,12 @@ class TutorielController extends Controller
      */
     public function update(UpdateTutorielRequest $request, Tutoriel $tutoriel)
     {
+        $user = Auth::user();
+        if(!$user->can('edit_tutoriel'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $tutoriel->update($request->validated());
         return to_route('tutoriel.index')->with('status', 'Modification effectuée avec succès');
     }
@@ -68,6 +106,12 @@ class TutorielController extends Controller
      */
     public function destroy(Tutoriel $tutoriel)
     {
+        $user = Auth::user();
+        if(!$user->can('delete_tutoriel'))
+        {
+            return redirect(route('home'))->with('info', 'Vous n\'avez pas l\'autorisation requise pour accéder à cette ressource');
+        }
+
         $tutoriel->delete();
         return to_route('tutoriel.index')->with('status', 'suppression effectuée avec succès');
     }
